Further work, blocker on moving
Moving is not recorded in anyway in the current icon datastructure. That is kind of a problem. Similarly we dont record the order icons are pulled from a DMI.

Might put this aside and work on Icon itself, kind of hard to add features for things that dont even exist yet.